Extract models to a separate file
