+++ universal-login-sdk/lib/index.ts
@@ -3,7 +3,8 @@ export default UniversalLoginSDK;
 export {SdkConfig} from './config/SdkConfig';
 export {SdkSigner} from './api/SdkSigner';
 export {FutureWallet, BalanceDetails} from './api/FutureWalletFactory';
-export {WalletService, WalletStorage, SerializedWalletState} from './core/services/WalletService';
+export {WalletService} from './core/services/WalletService';
 export {TransferService, encodeTransfer} from './core/services/TransferService';
 export {DeployedWallet} from './api/DeployedWallet';
 export {setBetaNotice} from './core/utils/setBetaNotice';
+export {WalletStorage, SerializedWalletState} from './core/models/WalletService';
